Working on fixing problems with first derivatives.
The first time derivatives of fields are now mostly computable if
solving time dpenedent equations.  It used to be that the first
derivatives were only computed if a field's second derivative was
in the equations (ie, displacement when a mass density property is
present).  If a system had viscoelasticity but no mass density,
the equations would be solved correctly but the time derivative
couldn't be computed as an output.

Now if the first derivative isn't computed during the solution,
it's automatically computed afterwards for all time dependent
problems.  It's not working yet for the time derivatives of the
out-of-plane components or for time-dependent fields at nodes where
time dependent Dirichlet boundary conditions are applied.

This isn't really ready to be committed but since I'll be away for
the next two weeks it seems wise to have a copy somewhere.  There
are some false starts and  debugging code  that will have to be
removed later.