Prefer the IC preconditioner over ILU or ILUT for ConjugateGradient.
In Eigen, ILU sometimes doesn't work well with CG.   I think its
because ILU doesn't preserve the matrix symmetry.  All test scripts
and reference files that use CG now use the IC preconditioner
instead, as does the Relax SkeletonModifier, and the symmetric
solver in Basic solver mode.

The way that OOF2 MatrixMethod/Preconditioner pairs are translated
to Eigen solver/preconditioner templated methods is now simpler.
Instead of a single solver_map dictionary in matrixmethod.py, each
PreconditionedMatrixMethod subclass has its own solver_map.  The
keys to the map are just the preconditioner classes themselves,
instead of the strings that were used before (and which weren't the
same as the class names).

There is now a PreconditionerParameter class with its own
PreconditionerWidget, which looks for the MatrixMethod widget and
only lists the Preconditioners that are compatible with it (ie,
listed in its solver_map).

The CG/ILU combination isn't currently prohibited, but it should
at least raise a warning.