Check that asymmetric_solver is not None before using it.
Also added an asymmetric_solver arg to many Set_Solver calls in
tests.  The GUI doesn't allow a user to avoid choosing an asymmetric
solver, but some very old scripts didn't have the argument.  If the
solver isn't used, it's ok for it to be None, but time dependent
problems might need it to solve for the first derivatives.