Still working on time derivative field and static_flux_offset.
Refactoring Property::flux_matrix() in the base class so that it
can take the numerical derivative with respect to the time derivative
fields as well.  This version doesn't do that yet, but the new code
computes the K matrix correctly.

Changed Property::flux_offset() in the base class so that it
can compute the offset numerically when flux_value() or
static_flux_value() is defined in the derived class.